Begin to split up tests by their scope
