+++ packages/arbor-store/tests/arbor.test.ts
@@ -1,15 +1,10 @@
 import { describe, expect, it, vi } from "vitest"
 import { Arbor } from "../src/arbor"
 import { ArborProxiable, detached, proxiable } from "../src/decorators"
-import {
-  ArborError,
-  DetachedNodeError,
-  NotAnArborNodeError,
-} from "../src/errors"
+import { DetachedNodeError } from "../src/errors"
 import { Seed } from "../src/path"
 import { Path } from "../src/path/path"
-import { ArborNode } from "../src/types"
-import { detach, isDetached, merge, pathFor, unwrap } from "../src/utilities"
+import { detach, pathFor, unwrap } from "../src/utilities"
 
 import { isNode } from "../src/guards"
 import { ScopedStore } from "../src/scoping/store"
@@ -1745,379 +1740,6 @@ describe("Arbor", () => {
 
       expect(store.state.get("abc")!.text).toBe("House cleaned")
     })
-  })
-
-  describe("Example: Utility functions", () => {
-    describe("detach", () => {
-      it("cannot detach the state tree's root node", () => {
-        const store = new Arbor({
-          todos: [
-            { id: 1, text: "Do the dishes" },
-            { id: 2, text: "Walk the dogs" },
-          ],
-        })
-
-        expect(() => detach(store.state)).toThrow(ArborError)
-      })
-
-      it("cannot detach values that are not already attached to the state tree", () => {
-        expect(() => detach(123)).toThrow(NotAnArborNodeError)
-        expect(() => detach("")).toThrow(NotAnArborNodeError)
-        expect(() => detach({})).toThrow(NotAnArborNodeError)
-      })
-
-      it("cannot detach node already detached", () => {
-        const store = new Arbor({
-          todos: [
-            { id: 1, text: "Do the dishes" },
-            { id: 2, text: "Walk the dogs" },
-          ],
-        })
-
-        const node = store.state.todos[0]
-        detach(node)
-
-        expect(() => detach(node)).toThrow(DetachedNodeError)
-      })
-
-      it("detaches a given ArborNode from the state tree", () => {
-        const initialState = {
-          todos: [
-            { id: 1, text: "Do the dishes" },
-            { id: 2, text: "Walk the dogs" },
-          ],
-        }
-
-        const todo0 = initialState.todos[0]
-
-        const store = new Arbor(initialState)
-
-        const detched = detach(store.state.todos[0])
-
-        expect(detched).toBe(todo0)
-        expect(store.state).toEqual({
-          todos: [{ id: 2, text: "Walk the dogs" }],
-        })
-      })
-
-      it("detaches children from a Map node", () => {
-        const todos = new Map<string, ArborNode<{ text: string }>>()
-        todos.set("a", { text: "Do the dishes" })
-        todos.set("b", { text: "Clean the house" })
-
-        const store = new Arbor(todos)
-
-        const todo = store.state.get("b")!
-        const detachedTodo = detach(todo)
-
-        expect(todos.get("b")).toBeUndefined()
-        expect(store.state.get("b")).toBeUndefined()
-        expect(detachedTodo).toBe(unwrap(todo))
-      })
-    })
-
-    describe("merge", () => {
-      it("cannot merge values that are not already attached to the state tree", () => {
-        const node = { name: "Alice", age: 32 }
-
-        expect(() => {
-          merge(node, { name: "Alice Doe", age: 33 })
-        }).toThrow(NotAnArborNodeError)
-      })
-
-      it("cannot merge data into node when node is detached from the state tree", () => {
-        const store = new Arbor({
-          todos: [
-            { id: 1, text: "Do the dishes" },
-            { id: 2, text: "Walk the dogs" },
-          ],
-        })
-
-        const node = store.state.todos[0]
-        detach(node)
-
-        expect(() => {
-          merge(node, { text: "" })
-        }).toThrow(DetachedNodeError)
-      })
-
-      it("merges data into a given state tree node", () => {
-        const store = new Arbor({
-          todos: [
-            { id: 1, text: "Do the dishes", active: false },
-            { id: 2, text: "Walk the dogs", active: true },
-          ],
-        })
-
-        const updated = merge(store.state.todos[0], {
-          text: "Did the dishes",
-          active: false,
-        })
-
-        expect(updated).toBe(store.state.todos[0])
-        expect(store.state).toEqual({
-          todos: [
-            { id: 1, text: "Did the dishes", active: false },
-            { id: 2, text: "Walk the dogs", active: true },
-          ],
-        })
-      })
-    })
-
-    describe("pathFor", () => {
-      it("cannot determine a path for a value that is not attached to the state tree", () => {
-        const node = { name: "Alice", age: 32 }
-
-        expect(() => {
-          pathFor(node)
-        }).toThrow(NotAnArborNodeError)
-      })
-
-      it("cannot determine a path for a detached node", () => {
-        const store = new Arbor({
-          todos: [
-            { id: 1, text: "Do the dishes" },
-            { id: 2, text: "Walk the dogs" },
-          ],
-        })
-
-        const node = store.state.todos[0]
-        detach(node)
-
-        expect(() => {
-          pathFor(node)
-        }).toThrow(DetachedNodeError)
-      })
-
-      it("determines the path of the node within the state tree", () => {
-        const store = new Arbor({
-          todos: [
-            { id: 1, text: "Do the dishes" },
-            { id: 2, text: "Walk the dogs" },
-          ],
-        })
-
-        const rootPath = Path.root(Seed.from(store.state))
-        const todosPath = rootPath.child(Seed.from(store.state.todos))
-        const todo0Path = todosPath.child(Seed.from(store.state.todos[0]))
-        const todo1Path = todosPath.child(Seed.from(store.state.todos[1]))
-
-        expect(pathFor(store.state)).toEqual(rootPath)
-        expect(pathFor(store.state.todos)).toEqual(todosPath)
-        expect(pathFor(store.state.todos[0])).toEqual(todo0Path)
-        expect(pathFor(store.state.todos[1])).toEqual(todo1Path)
-      })
-    })
-
-    describe("isDetached", () => {
-      it("returns true if value is not an Arbor node", () => {
-        const node = { name: "Alice", age: 32 }
-
-        expect(isDetached(node)).toBe(true)
-      })
-
-      it("determines if a node is no longer within the state tree", () => {
-        const store = new Arbor({
-          todos: [
-            { id: 1, text: "Do the dishes" },
-            { id: 2, text: "Walk the dogs" },
-          ],
-        })
-
-        const node = store.state.todos[0]
-
-        expect(isDetached(node)).toBe(false)
-
-        detach(node)
-
-        expect(isDetached(node)).toBe(true)
-      })
-    })
-
-    describe("unwrap", () => {
-      it("throws an error if given value is not a state tree node", () => {
-        const node = { name: "Alice", age: 32 }
-
-        expect(() => {
-          unwrap(node)
-        }).toThrow(NotAnArborNodeError)
-      })
-
-      it("returns the value wrapped by the node", () => {
-        const initialState = {
-          todos: [
-            { id: 1, text: "Do the dishes" },
-            { id: 2, text: "Walk the dogs" },
-          ],
-        }
-
-        const todo0 = initialState.todos[0]
-        const store = new Arbor(initialState)
-
-        expect(unwrap(store.state.todos[0])).toBe(todo0)
-      })
-    })
-  })
-
-  describe("path tracking", () => {
-    it("leverages structural sharing to preserve identities of nodes in the state tree", () => {
-      const store = new Arbor({
-        todos: [
-          { id: 1, text: "Do the dishes", active: false },
-          { id: 2, text: "Walk the dogs", active: true },
-        ],
-      })
-
-      const scopedStore = new ScopedStore(store)
-      const root = scopedStore.state
-      const todos = scopedStore.state.todos
-      const todo0 = scopedStore.state.todos[0]
-      const todo1 = scopedStore.state.todos[1]
-
-      todo0.active = true
-
-      expect(root).not.toBe(scopedStore.state)
-      expect(todos).not.toBe(scopedStore.state.todos)
-      expect(todo0).not.toBe(scopedStore.state.todos[0])
-      expect(todo1).toBe(scopedStore.state.todos[1])
-
-      expect(root).toEqual(scopedStore.state)
-      expect(todos).toEqual(scopedStore.state.todos)
-      expect(todo0).toEqual(scopedStore.state.todos[0])
-    })
-
-    it("reacts to mutations targeting paths being tracked", () => {
-      const store = new Arbor({
-        todos: [
-          { id: 1, text: "Do the dishes", active: false },
-          { id: 2, text: "Walk the dogs", active: true },
-        ],
-      })
-
-      const subscriber = vi.fn()
-      const scopedStore = new ScopedStore(store)
-
-      scopedStore.subscribe(subscriber)
-
-      // Tracks the following store paths
-      //  1. scopedStore.state.todos
-      //  2. scopedStore.state.todos[0]
-      //  2. scopedStore.state.todos[0].active
-      scopedStore.state.todos[0].active
-
-      // Does not notifies scopedStore subscribers
-      store.state.todos[0].id = 3
-      store.state.todos[0].text = "Do the dishes again"
-      // Does not notifies scopedStore subscribers
-      store.state.todos[1].id = 4
-      store.state.todos[1].text = "Walk the dogs again"
-      store.state.todos[1].active = false
-      store.state.todos[1] = { id: 3, text: "Clean the house", active: false }
-
-      expect(subscriber).not.toHaveBeenCalled()
-
-      store.state.todos[0].active = true
-      store.state.todos[0] = { id: 3, text: "Clean the house", active: false }
-      store.state.todos = []
-
-      expect(subscriber).toHaveBeenCalledTimes(3)
-    })
-
-    it("reacts to mutations targeting the root of the state tree", () => {
-      const store = new Arbor({
-        todos: [
-          { id: 1, text: "Do the dishes", active: false },
-          { id: 2, text: "Walk the dogs", active: true },
-        ],
-      })
-
-      const subscriber = vi.fn()
-      const scopedStore = new ScopedStore(store)
-
-      scopedStore.subscribe(subscriber)
-
-      store.setState({ todos: [] })
-
-      expect(subscriber).toHaveBeenCalledTimes(1)
-    })
-
-    it("marks nodes as tracked", () => {
-      const store = new Arbor({
-        todos: [
-          { id: 1, text: "Do the dishes", active: false },
-          { id: 2, text: "Walk the dogs", active: true },
-        ],
-      })
-
-      const scopedStore1 = new ScopedStore(store)
-
-      expect(scopedStore1.state).toBeTrackedNode()
-      expect(scopedStore1.state.todos).toBeTrackedNode()
-      expect(scopedStore1.state.todos[0]).toBeTrackedNode()
-      expect(scopedStore1.state.todos[1]).toBeTrackedNode()
-    })
-
-    it("automatically unwraps tracked node when creating a derived tracking scope", () => {
-      const store = new Arbor({
-        todos: [
-          { id: 1, text: "Do the dishes", active: false },
-          { id: 2, text: "Walk the dogs", active: true },
-        ],
-      })
-
-      const scopedStore1 = new ScopedStore(store)
-
-      expect(scopedStore1.state).toEqual(store.state)
-      expect(scopedStore1.state).not.toBe(store.state)
-      expect(unwrap(scopedStore1.state)).toBe(store.state)
-
-      const scopedStore2 = new ScopedStore(scopedStore1.state.todos[0])
-
-      expect(scopedStore2.state).toEqual(store.state.todos[0])
-      expect(scopedStore2.state).not.toBe(store.state.todos[0])
-      expect(scopedStore2.state).not.toBe(scopedStore1.state.todos[0])
-      expect(unwrap(scopedStore2.state)).toBe(store.state.todos[0])
-    })
-
-    it("isolates subscriptions to their own tracking scope", () => {
-      const store = new Arbor({
-        todos: [
-          { id: 1, text: "Do the dishes", active: false },
-          { id: 2, text: "Walk the dogs", active: true },
-        ],
-      })
-
-      const subscriber1 = vi.fn()
-      const subscriber2 = vi.fn()
-
-      const scopedStore1 = new ScopedStore(store)
-      scopedStore1.subscribe(subscriber1)
-
-      // Causes scopedStore1 to watch mutations to the following paths:
-      // 1. root
-      // 2. root.todos
-      // 3. root.todos[0]
-      const firstTodo = scopedStore1.state.todos[0]
-
-      const scopedStore2 = new ScopedStore(firstTodo)
-      scopedStore2.subscribe(subscriber2)
-
-      // Causes scopedStore2 to watch mutations to the following paths but does not affect scopedStore1:
-      // 1. root.todos[0].active
-      scopedStore2.state.active = true
-
-      expect(subscriber1).not.toHaveBeenCalled()
-      expect(subscriber2).toHaveBeenCalledTimes(1)
-
-      scopedStore1.state.todos[2] = {
-        id: 3,
-        text: "Learn Arbor",
-        active: true,
-      }
-
-      expect(subscriber1).toHaveBeenCalledTimes(1)
-      expect(subscriber2).toHaveBeenCalledTimes(1)
-    })
 
     it("allows pushing the same item multiple times to the array", () => {
       const store = new Arbor([{ name: "Alice" }, { name: "Carol" }])
@@ -2139,352 +1761,34 @@ describe("Arbor", () => {
       ])
     })
 
-    it("is able to track nodes resulted from method calls", () => {
-      const store = new Arbor({
-        todos: [
-          { id: 1, text: "Do the dishes", active: false },
-          { id: 2, text: "Walk the dogs", active: true },
-        ],
-      })
-
-      const scopedStore = new ScopedStore(store)
-
-      const activeTodo = scopedStore.state.todos.find((t) => t.active)
-
-      expect(activeTodo).toBeTrackedNode()
-      expect(activeTodo).toBe(scopedStore.state.todos[1])
-    })
-
-    it("handles mutations to the root of the store", () => {
-      const store = new Arbor([
-        { id: 1, text: "Do the dishes", active: false },
-        { id: 2, text: "Walk the dogs", active: true },
-      ])
-
-      const subscriber = vi.fn()
-      const scopedStore = new ScopedStore(store)
-      scopedStore.subscribe(subscriber)
-
-      scopedStore.state.push({ id: 3, text: "Walk the dogs", active: true })
-
-      expect(subscriber).toHaveBeenCalledTimes(1)
-    })
-
-    it("allows tracking a specific node within the state tree", () => {
-      const store = new Arbor([
-        { id: 1, text: "Do the dishes", active: false },
-        { id: 2, text: "Walk the dogs", active: true },
-      ])
-
-      const scopedStore = new ScopedStore(store.state[0])
-
-      expect(unwrap(scopedStore.state)).toBe(store.state[0])
-    })
-
-    it("creates a virtual state tree from a given subtree", () => {
-      const store = new Arbor([
-        { id: 1, text: "Do the dishes", active: false },
-        { id: 2, text: "Walk the dogs", active: true },
-      ])
-
-      const scopedStore = new ScopedStore(store.state[0])
-
-      expect(unwrap(scopedStore.state)).toBe(store.state[0])
-
-      const newState = scopedStore.setState({
-        id: 3,
-        text: "Learn Arbor",
-        active: true,
-      })
-
-      expect(unwrap(newState)).toBe(store.state[0])
-      expect(newState).toEqual({
-        id: 3,
-        text: "Learn Arbor",
-        active: true,
-      })
-    })
-
-    it("tracks new array items being added", () => {
-      const store = new Arbor([{ name: "Alice" }, { name: "Bob" }])
-
-      const subscriber = vi.fn()
-      const scopedStore = new ScopedStore(store)
-      scopedStore.subscribe(subscriber)
-
-      store.state.push({ name: "Carol" })
-
-      expect(subscriber).toHaveBeenCalledTimes(1)
-    })
-
-    it("tracks new props being added to nodes", () => {
-      const store = new Arbor({ users: [{ name: "Alice" }, { name: "Bob" }] })
-
-      const subscriber = vi.fn()
-      const scopedStore = new ScopedStore(store)
-      scopedStore.subscribe(subscriber)
-
-      // track state.users
-      scopedStore.state.users[1]
-
-      store.state.users[2] = { name: "Carol" }
-
-      expect(subscriber).toHaveBeenCalledTimes(1)
-    })
-
-    it("preserves children node path tracking when plucking them through a method operation (like filter, map, etc...)", () => {
-      const store = new Arbor([
-        { name: "Carol", active: true },
-        { name: "Alice", active: false },
-      ])
-      const tracked = new ScopedStore(store.state)
-
-      store.state.filter // binds filter to the original store
-      const filterBoundToTrackedStore = tracked.state.filter
-      const activeUsers = filterBoundToTrackedStore((u) => u.active)
-
-      expect(activeUsers[0]).toBeTrackedNode()
-    })
-
-    it("preserves path tracking on nodes 'plucked' from the state tree", () => {
-      const store = new Arbor([
-        { name: "Carol", active: true },
-        { name: "Alice", active: false },
-      ])
-      const tracked = new ScopedStore(store.state)
-      const subscriber = vi.fn()
-
-      tracked.subscribe(subscriber)
-
-      const carol = tracked.state[0]
-
-      carol.active = false
-
-      expect(carol).toBeTrackedNode()
-      expect(subscriber).toHaveBeenCalledTimes(1)
-    })
-
-    it("can track path access on nodes 'pluck' from the state tree", () => {
-      const store = new Arbor({
-        users: [
-          { name: "Carol", active: true },
-          { name: "Alice", active: false },
-        ],
-      })
-      const tracked = new ScopedStore(store)
-
-      const carol = tracked.state.users[0]
-      expect(carol).toBeTrackedNode()
-    })
-
-    it("ensure node methods have stable memory reference across updates", () => {
-      const store = new Arbor({
-        todos: ["Do the dishes"],
-        users: ["Carol", "Alice"],
-      })
-
-      const tracked = new ScopedStore(store)
-      const userFilter = tracked.state.users.filter
-
-      expect(userFilter).toBe(tracked.state.users.filter)
-
-      tracked.state.todos[0] = "Clean the house"
-
-      expect(userFilter).toBe(tracked.state.users.filter)
-
-      tracked.state.users.push("Bob")
-
-      expect(userFilter).toBe(tracked.state.users.filter)
-    })
-
-    it("can safely use a method reference across node updates", () => {
-      const store = new Arbor({
-        users: [
-          { name: "Carol", active: true },
-          { name: "Alice", active: false },
-        ],
-      })
-
-      const tracked = new ScopedStore(store)
-      const userFilter = tracked.state.users.filter
-
-      expect(userFilter((u) => u.active)).toEqual([
-        { name: "Carol", active: true },
-      ])
-
-      store.state.users[1].active = true
-
-      expect(userFilter((u) => u.active)).toEqual([
-        { name: "Carol", active: true },
-        { name: "Alice", active: true },
-      ])
-
-      store.state.users[0].active = false
-
-      expect(userFilter((u) => u.active)).toEqual([
-        { name: "Alice", active: true },
-      ])
-    })
-
-    it("can handle null props", () => {
-      const store = new Arbor({ name: "Carol", email: null })
-      const tracked = new ScopedStore(store.state)
-
-      expect(tracked.state.email).toBeNull()
-    })
-  })
-
-  it("does not track detached props", () => {
-    @proxiable
-    class SomeNode {
-      @detached untrackedProp = "untracked"
-      trackedProp = "tracked"
-    }
-
-    const store = new ScopedStore(new Arbor(new SomeNode()))
-
-    store.state.untrackedProp
-    store.state.trackedProp
-
-    expect(store.scope.isTracking(store.state, "trackedProp")).toBe(true)
-    expect(store.scope.isTracking(store.state, "untrackedProp")).toBe(false)
-  })
-
-  it("binds methods to the path tracking proxy", () => {
-    @proxiable
-    class Todo {
-      constructor(public text = "") {}
-    }
-
-    @proxiable
-    class TodoApp {
-      todos: Todo[] = []
-
-      removeTodo(todo) {
-        this.todos = this.todos.filter((t) => t !== todo)
-      }
-    }
-
-    const store = new ScopedStore(new Arbor(new TodoApp()))
-    const subscriber = vi.fn()
-
-    store.subscribe(subscriber)
-
-    const state = store.state
-    state.todos = [new Todo("Do the dishes"), new Todo("Clean the house")]
-    state.removeTodo(state.todos[0])
-    state.todos.push(new Todo("Walk the dogs"))
-
-    expect(subscriber.mock.calls.length).toBe(3)
-    expect(subscriber.mock.calls[0][0].metadata.operation).toEqual("set")
-    expect(subscriber.mock.calls[1][0].metadata.operation).toEqual("set")
-    expect(subscriber.mock.calls[2][0].metadata.operation).toEqual("push")
-  })
-
-  it("binds methods to the instance of the node they belong to", () => {
-    @proxiable
-    class Todo {
-      constructor(public text = "", public done = false) {}
-
-      toggle() {
-        this.done = !this.done
-      }
-    }
-
-    @proxiable
-    class TodoApp {
-      todos: Todo[] = []
-    }
-
-    const store = new ScopedStore(new Arbor(new TodoApp()))
-    const subscriber = vi.fn()
-
-    store.subscribe(subscriber)
-
-    const state = store.state
-    state.todos = [new Todo("Do the dishes"), new Todo("Clean the house")]
-
-    state.todos[0].toggle()
-    state.todos[1].toggle()
-
-    expect(state.todos[0].done).toBe(true)
-    expect(state.todos[1].done).toBe(true)
-  })
-
-  it("prevents object methods from conflicting with the Proxy API", () => {
-    @proxiable
-    class NumberList {
-      items = []
-
-      get(index: number) {
-        return this.items[index]
+    it("prevents object methods from conflicting with the Proxy API", () => {
+      @proxiable
+      class NumberList {
+        items: number[] = []
+
+        get(index: number) {
+          return this.items[index]
+        }
+
+        set(index: number, value: number) {
+          this.items[index] = value
+        }
+
+        deleteProperty(index: number) {
+          this.items.splice(index, 1)
+        }
       }
 
-      set(index: number, value: number) {
-        this.items[index] = value
-      }
+      const todos = new Arbor(new NumberList())
 
-      deleteProperty(index: number) {
-        this.items.splice(index, 1)
-      }
-    }
-
-    const todos = new Arbor(new NumberList())
-
-    todos.state.set(0, 1)
-    todos.state.set(1, 2)
-    expect(todos.state.get(0)).toBe(1)
-    expect(todos.state.get(1)).toBe(2)
-    todos.state.deleteProperty(0)
-    expect(todos.state.get(0)).toBe(2)
-    expect(todos.state.get(1)).toBeUndefined()
-  })
-
-  it("does not proxy instances of Arbor when used as fields of a store", () => {
-    const store1 = new Arbor<Array<{ text: string }>>([
-      { text: "Do the dishes" },
-    ])
-
-    @proxiable
-    class Form {
-      constructor(readonly store = store1) {}
-
-      addTodo(text: string) {
-        this.store.state.push({ text })
-      }
-    }
-
-    const store2 = new ScopedStore(new Arbor(new Form(store1)))
-
-    store2.state.addTodo("LOL")
-
-    expect(store1.state).toEqual([{ text: "Do the dishes" }, { text: "LOL" }])
-  })
-
-  it("track nodes accessed through a getter", () => {
-    @proxiable
-    class TodoApp {
-      todos = [
-        { text: "Do the dishes", done: false },
-        { text: "Walk the dogs", done: true },
-      ]
-
-      get activeTodos() {
-        return this.todos.filter((todo) => !todo.done)
-      }
-    }
-
-    const scoped = new ScopedStore(new Arbor(new TodoApp()))
-    const todo = scoped.state.activeTodos[0]
-
-    expect(scoped).toBeTracking(todo, "done")
-    expect(scoped).not.toBeTracking(todo, "text")
-
-    // access the first todo in the result so the scoped store
-    // start tracking changes to the todo's "text" prop.
-    todo.text
-
-    expect(scoped).toBeTracking(todo, "text")
+      todos.state.set(0, 1)
+      todos.state.set(1, 2)
+      expect(todos.state.get(0)).toBe(1)
+      expect(todos.state.get(1)).toBe(2)
+      todos.state.deleteProperty(0)
+      expect(todos.state.get(0)).toBe(2)
+      expect(todos.state.get(1)).toBeUndefined()
+    })
   })
 
   describe("state tree diff", () => {
